requirements.txt
django
whitenoise
python-dotenv
openai
azure-identity
django-htmx